requirements.txt
streamlit
librosa
soundfile
tensorflow
scikit-learn
google-generativeai
python-dotenv
numpy
